requirements.txt

Django==3.2
gunicorn==20.1.0
mysqlclient==2.0.3
django-allauth==0.42.0
